--- a/benchmark/benchmark-normalization.hs
+++ b/benchmark/benchmark-normalization.hs
@@ -4,6 +4,8 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 {-# OPTIONS_GHC -Wno-partial-type-signatures #-}
 
+import           Clash.Signal (VDomainConfiguration)
+
 import           Clash.Annotations.BitRepresentation.Internal (CustomReprs)
 import           Clash.Core.TyCon
 import           Clash.Core.Var
@@ -20,8 +22,10 @@ import           Criterion.Main
 
 import qualified Control.Concurrent.Supply    as Supply
 import           Control.DeepSeq              (NFData(..), rwhnf)
+import           Data.HashMap.Strict          (HashMap)
 import           Data.IntMap.Strict           (IntMap)
 import           Data.List                    (isPrefixOf, partition)
+import           Data.Text                    (Text)
 import           System.Environment           (getArgs, withArgs)
 
 import BenchmarkCommon
@@ -46,7 +50,7 @@ main = do
 benchFile :: [FilePath] -> FilePath -> Benchmark
 benchFile idirs src =
   env (setupEnv idirs src) $
-    \ ~((bindingsMap,tcm,tupTcm,_topEntities,primMap,reprs,topEntityNames,topEntity),supplyN) -> do
+    \ ~((bindingsMap,tcm,tupTcm,_topEntities,primMap,reprs,_domainConfs,topEntityNames,topEntity),supplyN) -> do
       bench ("normalization of " ++ src)
             (nf (normalizeEntity reprs bindingsMap primMap tcm tupTcm typeTrans
                                  ghcEvaluator
@@ -59,7 +63,7 @@ setupEnv
   -> FilePath
   -> IO ((BindingMap, TyConMap, IntMap TyConName
          ,[TopEntityT]
-         ,CompiledPrimMap, CustomReprs, [Id], Id
+         ,CompiledPrimMap, CustomReprs, HashMap Text VDomainConfiguration, [Id], Id
          )
         ,Supply.Supply
         )