--- a/benchmark/clash-benchmark.cabal
+++ b/benchmark/clash-benchmark.cabal
@@ -19,6 +19,7 @@ library
                        containers,
                        ghc,
                        mtl,
+                       text,
                        unordered-containers,
 
                        clash-ghc,
@@ -36,9 +37,22 @@ executable clash-benchmark-normalization
                        deepseq,
                        directory,
                        filepath,
+                       text,
                        unordered-containers,
 
                        clash-benchmark,
                        clash-ghc,
                        clash-lib,
                        clash-prelude
+
+executable clash-benchmark-concurrency
+  main-is:             benchmark-concurrency.hs
+  default-language:    Haskell2010
+  ghc-options:         -O2 -Wall -Wcompat -threaded -rtsopts -with-rtsopts=-N
+  build-depends:       base,
+                       criterion,
+                       time,
+
+                       clash-lib,
+                       clash-ghc,
+                       clash-benchmark