--- a/clash-ghc/src-bin-881/Clash/GHCi/UI.hs
+++ b/clash-ghc/src-bin-881/Clash/GHCi/UI.hs
@@ -2171,7 +2171,7 @@ makeHDL backend startAction optsRef srcs = do
                   topEntities
                   mainTopEntity
                   opts2
-                  (startTime,prepTime)
+                  startTime
 
 makeVHDL :: IORef ClashOpts -> [FilePath] -> InputT GHCi ()
 makeVHDL = makeHDL' (Clash.Backend.initBackend @VHDLState)