--- a/clash-ghc/src-bin-9.0/Clash/GHCi/UI.hs
+++ b/clash-ghc/src-bin-9.0/Clash/GHCi/UI.hs
@@ -2288,7 +2288,7 @@ makeHDL backend startAction optsRef srcs = do
                   topEntities
                   mainTopEntity
                   opts2
-                  (startTime,prepTime)
+                  startTime
 
 makeVHDL :: IORef ClashOpts -> [FilePath] -> InputT GHCi ()
 makeVHDL = makeHDL' (Clash.Backend.initBackend @VHDLState)