… benchmark/common/BenchmarkCommon.hs
… benchmark/profiling/prepare/profile-normalization-prepare.hs
… benchmark/profiling/run/profile-netlist-run.hs
… benchmark/tests/ManyEntitiesEqual.hs
… benchmark/tests/ManyEntitiesVaried.hs
… changelog/2022-01-14T09_39_07+01_00_async_at_synthesis_boundaries.md
… tests/src/Test/Tasty/Clash/NetlistTest.hs