@martijnbastiaan 2022-01-13
Related issue: https://github.com/clash-lang/clash-compiler/issues/354